pyproject.toml
[project]
name = "storyblok-mcp-server"
version = "0.1.0"
description = "MCP server for storyblok"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
